requirements.txt
SpeechRecognition==3.10.1
pydub==0.25.1
pocketsphinx==5.0.3
transformers==4.36.2
faiss-cpu==1.7.4
torch==2.1.2
openai==1.6.1
typer==0.9.0
pandas==2.1.4
loguru==0.7.2
typing-extensions==4.9.0
tabulate==0.9.0
